+++ web/documentserver-example/php-laravel/app/Http/Controllers/EditorController.php
@@ -245,6 +245,7 @@ class EditorController extends Controller
 
         if ($user['id'] != 'uid-0') {
             $historyLayout .= "// add mentions for not anonymous users
+                config.events['onRequestRefreshFile'] = onRequestRefreshFile;
                 config.events['onRequestClose'] = onRequestClose;
                 config.events['onRequestUsers'] = onRequestUsers;
                 config.events['onRequestSaveAs'] = onRequestSaveAs;