+++ web/documentserver-example/php-laravel/app/Repositories/UserRepository.php
@@ -75,6 +75,7 @@ class UserRepository
             "Can't view chat",
             'View file without collaboration',
             'Can’t submit forms',
+            'Can’t refresh outdated file',
         ];
         $this->users = [
             new User(